fix(history): stop field history printing the server's own map syntax
A change to a jsonb column rendered through Go's default formatting, so
the account page showed `map[logo:https://scale.sc/apple-touch-icon.png]`
where a value belonged. Scalars now render as themselves and structured
values as JSON. A value that cannot be encoded at all says so in words
rather than falling back to Go's rendering of it.

The existing test asserted only that a structured value rendered
non-empty, which is why the leak survived it; it now pins the shape.

Two things found by reading the page rather than the code:

The timeline's Changes rows collided on React keys. `FieldHistoryEntry.id`
is the AUDIT row's id and one audit row projects one entry per field it
touched, so the id repeats across a flat list — only the (id, field) pair
identifies a row. The per-field view never saw this because its groups
hold one field each.

Either side of a diff can be a whole jsonb document — one scraped fact
set here runs past ten thousand characters — so each side is capped at a
few lines and scrolls. The full value stays reachable; it just cannot
push the rest of the history off the screen.

fieldhistory.go crossed the 500-line cap, so the projection of one audit
row into per-field changes — mask, diff, value rendering — moved to
fieldhistorydiff.go. That file never touches the database; the reads
that find and page the rows stay put.
